refactor: refacotr extensions
