+++ Mongsil/Mongsil/Module/Component/Extensions/View/View+extensions.swift
@@ -169,20 +169,3 @@ public struct AdaptsToKeyboard: ViewModifier {
     }
   }
 }
-
-extension String {
-  func textToImage() -> UIImage? {
-    let nsString = (self as NSString)
-    let font = UIFont.systemFont(ofSize: 40)
-    let stringAttributes = [NSAttributedString.Key.font: font]
-    let imageSize = nsString.size(withAttributes: stringAttributes)
-
-    UIGraphicsBeginImageContextWithOptions(imageSize, false, 0)
-    UIColor.clear.set()
-    UIRectFill(CGRect(origin: CGPoint(), size: imageSize))
-    nsString.draw(at: CGPoint.zero, withAttributes: stringAttributes)
-    let image = UIGraphicsGetImageFromCurrentImageContext()
-    UIGraphicsEndImageContext()
-    return image ?? UIImage()
-  }
-}
